requirements.txt
# Core Bot & Telegram
python-telegram-bot==20.7

# OpenAI API
openai==1.93.0

# FastAPI (untuk backend API jika kamu pakai)
fastapi==0.109.0
uvicorn==0.27.0

# Database & ORM
sqlalchemy==2.0.25
pymysql==1.1.0

# Env Handling & Logging
python-dotenv==1.0.1
bcrypt==4.1.2
